The url to get the json with the programs has been changed to: r'^ api /1/radiocom/programs and has been placed in the commons/url.py class
The programmes_json method has been moved to api/views.py.
Also the license was added to those classes and the imports were first placed before the ones from.
